refactor

+++ src/main/scala/org/sisioh/sip/message/header/CSeqHeader.scala
@@ -3,7 +3,7 @@ package org.sisioh.sip.message.header
 trait CSeqHeader extends Header {
 
   val method: String
-  val sequenceNumber: Int
+  val sequenceNumber: Long
 
 }
